# lib/api-zod/src/generated/types/addMemberBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { AddMemberBodyRole } from "./addMemberBodyRole";

export interface AddMemberBody {
  userId: number;
  role?: AddMemberBodyRole;
}