# lib/api-zod/src/generated/types/projectWithStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Project } from "./project";

export type ProjectWithStats = Project & {
  memberCount: number;
  taskCount: number;
  completedTaskCount: number;
  overdueTaskCount: number;
};