# lib/api-zod/src/generated/types/taskWithAssignee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Project } from "./project";
import type { Task } from "./task";
import type { User } from "./user";

export type TaskWithAssignee = Task & {
  assignee?: User | null;
  project: Project;
};